backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1
python-multipart==0.0.9
python-dotenv==1.0.1
aiofiles==23.2.1
sqlalchemy==2.0.25
alembic==1.13.1
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0
